boards: infineon: Add dual-core support for CY8CKIT-062S2-AI
Refactor dual-core support from separate board to board variants
within cy8ckit_062s2_ai. This allows the same board to support
M4-only, M0-only, and dual-core configurations using board
qualifiers.

Changes:
- Add board variants: m0, m4, and m4_dual qualifiers
- Add M0 and M4 DTS files with proper memory partitioning:
  * M0+: 512KB flash, 128KB SRAM
  * M4: 1536KB flash, 892KB SRAM
  * Shared: 4KB SRAM for inter-core communication
- Update SOC configuration to declare dual-core CPU clusters
- Move dual-core sample from cy8ckit_062s2_ai_dual_dual_core to
  cy8ckit_062s2_ai_dual_core
- Update board documentation with variant information and
  memory partitioning details
- Configure clock control: M0+ handles clock configuration
  in dual-core mode to avoid hardware conflicts

The board can now be used with:
- cy8ckit_062s2_ai/cy8c624abzi_s2d44/m4 (M4-only, default)
- cy8ckit_062s2_ai/cy8c624abzi_s2d44/m0 (M0-only)
- cy8ckit_062s2_ai/cy8c624abzi_s2d44/m0 + m4_dual (dual-core)

Partially addresses #97807
