+++ boards/infineon/cy8ckit_062s2_ai/Kconfig.cy8ckit_062s2_ai
@@ -2,4 +2,6 @@
 # SPDX-License-Identifier: Apache-2.0
 
 config BOARD_CY8CKIT_062S2_AI
-	select SOC_CY8C624ABZI_S2D44
+	select SOC_CY8C624ABZI_S2D44 if BOARD_CY8CKIT_062S2_AI_CY8C624ABZI_S2D44_M4
+	select SOC_CY8C624ABZI_S2D44_M4 if BOARD_CY8CKIT_062S2_AI_CY8C624ABZI_S2D44_M4_DUAL
+	select SOC_CY8C624ABZI_S2D44_M0 if BOARD_CY8CKIT_062S2_AI_CY8C624ABZI_S2D44_M0
