+++ boards/infineon/cy8ckit_062s2_ai/cy8ckit_062s2_ai_cy8c624abzi_s2d44_m0.yaml
@@ -2,16 +2,17 @@
 # an affiliate of Cypress Semiconductor Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-identifier: cy8ckit_062s2_ai_dual/cy8c624abzi_s2d44/m4
-name: CY8CKIT-062S2-AI Dual Core (M4)
+identifier: cy8ckit_062s2_ai/cy8c624abzi_s2d44/m0
+name: CY8CKIT-062S2-AI Dual Core (M0+)
 type: mcu
 arch: arm
-ram: 896
-flash: 1536
+ram: 128
+flash: 512
 toolchain:
   - zephyr
 supported:
   - gpio
   - uart
-  - watchdog
+  - wifi
+  - bluetooth
 vendor: infineon
