+++ boards/infineon/cy8ckit_062s2_ai/cy8ckit_062s2_ai_cy8c624abzi_s2d44_m0_defconfig
