+++ boards/infineon/cy8ckit_062s2_ai/cy8ckit_062s2_ai_cy8c624abzi_s2d44_m4_dual_defconfig
@@ -16,8 +16,5 @@ CONFIG_SERIAL=y
 # Enable GPIO driver
 CONFIG_GPIO=y
 
-# Enable clock controller
-# CONFIG_CLOCK_CONTROL=y
-
 # General configuration
 CONFIG_BUILD_OUTPUT_HEX=y
