--- boards/infineon/cy8ckit_062s2_ai_dual/Kconfig.cy8ckit_062s2_ai_dual
@@ -1,7 +1,0 @@
-# Copyright (c) 2025 Cypress Semiconductor Corporation (an Infineon company) or
-# an affiliate of Cypress Semiconductor Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-config BOARD_CY8CKIT_062S2_AI_DUAL
-	select SOC_CY8C624ABZI_S2D44_M4 if BOARD_CY8CKIT_062S2_AI_DUAL_CY8C624ABZI_S2D44_M4
-	select SOC_CY8C624ABZI_S2D44_M0 if BOARD_CY8CKIT_062S2_AI_DUAL_CY8C624ABZI_S2D44_M0
