--- boards/infineon/cy8ckit_062s2_ai_dual/board.cmake
@@ -1,9 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-# During gdb session, by default connect to CM4 core.
-board_runner_args(openocd "--gdb-init=disconnect")
-board_runner_args(openocd "--gdb-init=target extended-remote :3334")
-include(${ZEPHYR_BASE}/boards/common/openocd.board.cmake)
-
-board_runner_args(pyocd "--target=cy8c6xxa")
-include(${ZEPHYR_BASE}/boards/common/pyocd.board.cmake)
