--- boards/infineon/cy8ckit_062s2_ai_dual/board.yml
@@ -1,10 +1,0 @@
-# Copyright (c) 2025 Cypress Semiconductor Corporation (an Infineon company) or
-# an affiliate of Cypress Semiconductor Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-board:
-  name: cy8ckit_062s2_ai_dual
-  full_name: PSOC 6 AI Evaluation Kit (Dual Core)
-  vendor: infineon
-  socs:
-    - name: cy8c624abzi_s2d44
