--- boards/infineon/cy8ckit_062s2_ai_dual/support/openocd.cfg
@@ -1,12 +1,0 @@
-if {[info exists env(OPENOCD_INTERFACE)]} {
-        set INTERFACE $env(OPENOCD_INTERFACE)
-} else {
-        # By default connect over Debug USB port
-        set INTERFACE "cmsis-dap"
-}
-
-source [find interface/$INTERFACE.cfg]
-
-transport select swd
-
-source [find target/infineon/cy8cxxa.cfg]
