+++ samples/boards/infineon/cy8ckit_062s2_ai_dual_core/cm4/CMakeLists.txt
