+++ samples/boards/infineon/cy8ckit_062s2_ai_dual_core/sysbuild.cmake
@@ -1,5 +1,5 @@
 ExternalZephyrProject_Add(
   APPLICATION cm4
   SOURCE_DIR ${APP_DIR}/cm4
-  BOARD cy8ckit_062s2_ai_dual/cy8c624abzi_s2d44/m4
+  BOARD cy8ckit_062s2_ai/cy8c624abzi_s2d44/m4_dual
 )
